refactor(providers): migrate to vendor-based architecture
- Move authentication providers to vendors/supabase/auth/ structure
- Reorganize mock providers to vendors/mock/auth/ for consistency
- Simplify domain structure by moving core classes to domain/auth/
- Remove PROVIDER_SCOPES in favor of PROVIDER_VENDORS
- Update provider registries to use PROVIDER_DOMAINS constants
- Remove registry-keys.ts in favor of provider constants
- Maintain clean separation between vendor implementations and core interfaces
