+++ providers/domain/auth/auth-error.ts
