+++ providers/domain/auth/interfaces/auth-provider.interface.ts
