--- providers/domain/auth/mock/index.ts
@@ -1,1 +1,0 @@
-export { MockAuthProvider } from './mock-auth-provider';
