+++ providers/registry/index.ts
@@ -2,5 +2,4 @@ export {
   RegistryClientProvider,
   useClientRegistry,
 } from './registry-client-provider';
-export { REGISTRY_KEYS } from './registry-keys';
 export { serverProviders } from './registry-server-provider';
