+++ providers/registry/registry-server-provider.ts
@@ -4,11 +4,12 @@
  * This module should only be imported in server contexts.
  */
 
-import { SupabaseServerAuthProvider } from '../domain/auth/supabase/server-auth-provider';
+import { PROVIDER_DOMAINS } from '@/providers';
+import { SupabaseServerAuthProvider } from '../vendors/supabase/auth';
 
 // Export server providers directly - no registry needed
 export const serverProviders = {
-  auth: new SupabaseServerAuthProvider(),
+  [PROVIDER_DOMAINS.AUTH]: new SupabaseServerAuthProvider(),
   // Add future providers here
   // database: new ServerDatabaseProvider(),
   // storage: new ServerStorageProvider(),
