+++ providers/vendors/mock/index.ts
@@ -1,0 +1,1 @@
+export { MockAuthProvider } from './auth/mock-auth-provider';
